agent_sdk_v2/requirements.txt
# ELA Question Generation SDK v2
anthropic>=0.18.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0